requirements.txt
flask
python-dotenv
txtai[embeddings]
openai
gunicorn
